requirements.txt
-i https://pypi.org/simple
coverage==4.5.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-nose==1.4.6
django==2.2.2
djangorestframework==3.9.4
nose==1.3.7
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
